feat: add parse_range_header to http utils
