+++ Cargo.toml
@@ -84,3 +84,4 @@ tokio-rustls = "0.25"
 hyper-rustls = "0.26"
 http-body-util = "0.1.0"
 regex = "1.10.2"
+http-range-header = "0.4.0"
