+++ src/lib.rs
@@ -96,6 +96,10 @@ pub enum Error {
     #[error(transparent)]
     Elapsed(#[from] tokio_stream::Elapsed),
 
+    // RangeUnsatisfiableError is the error for range unsatisfiable.
+    #[error(transparent)]
+    RangeUnsatisfiableError(#[from] http_range_header::RangeUnsatisfiableError),
+
     // HashRing is the error for hashring.
     #[error{"hashring {0} is failed"}]
     HashRing(String),
